+++ src/controllers/users.controllers.js
@@ -12,7 +12,7 @@ class UsersController {
     try {
       const userId = res.locals.user.id;
       const userInfo = await this.usersService.userInfo(userId);
-      res.status(200).json({ data: userInfo });
+      res.status(this.code.ok.status).json({ data: userInfo });
     } catch {
       res.status(this.code.Forbidden.status).json(this.code.Forbidden.message);
     }
@@ -24,7 +24,9 @@ class UsersController {
       const userId = res.locals.user.id;
       const { userName, email, phone } = req.body;
       await this.usersService.updateUserInfo(userId, userName, email, phone);
-      res.status(200).json({ message: "회원 정보 수정이 완료되었습니다." });
+      res
+        .status(this.code.ok.status)
+        .json({ message: "회원 정보 수정이 완료되었습니다." });
     } catch {
       res.status(this.code.Forbidden.status).json(this.code.Forbidden.message);
     }
