+++ src/repository/auth.repository.js
@@ -17,12 +17,8 @@ class AuthRepository {
 
   // 로그인
   userLogin = async (email) => {
-    try{
-      const user = await User.findOne({ where: { email } });
-      return user
-    }
-    catch (err) {
-    }
+    const user = await User.findOne({ where: { email } });
+    return user;
   };
 }
